+++ src/components/TagsReadOnly.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import './TagsDisplay.css'
+import './TagsReadOnly.css'
 
 interface Props extends React.HTMLProps<HTMLDivElement> {
     tags: string[]
@@ -9,15 +9,15 @@ class component extends React.Component<Props> {
     render() {
         const { tags } = this.props
         return (
-            <div className={`cl-tags-display ${this.props.className}`}>
+            <div className={`cl-tags-readonly ${this.props.className}`}>
                 {tags.length === 0
                     ? `0 tags`
                     : tags.map((tag, i) =>
-                    <div className="cl-tags-display__tag" key={i}>
+                    <div className="cl-tags-readonly__tag" key={i}>
                         {tag}
                     </div>
                 )}
-                <span className="cl-tags-display__clearfix"></span>
+                <span className="cl-tags-readonly__clearfix"></span>
             </div>
         )
     }
